# backend/openapi/gones.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Gones.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": ""
    }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "LocalIdentityEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "LocalIdentityEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "LocalIdentityEndpoints"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/logout-all": {
      "post": {
        "tags": [
          "LocalIdentityEndpoints"
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/me": {
      "get": {
        "tags": [
          "LocalIdentityEndpoints"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "LocalIdentityEndpoints"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchUserProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/me/sessions/{id}": {
      "delete": {
        "tags": [
          "LocalIdentityEndpoints"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessTokenResponse": {
        "required": [
          "accessToken",
          "expiresAt",
          "tokenType"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "tokenType": {
            "type": "string"
          }
        }
      },
...
      "Instant": {},
      "LoginRequest": {
        "required": [
          "email",
          "password",
          "deviceLabel"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 0,
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "deviceLabel": {
            "maxLength": 100,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "rateLimitAccount": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PatchUserProfileRequest": {
        "required": [
          "username",
          "firstName",
          "lastName",
          "location",
          "birthYear",
          "preferredLanguage",
          "isFirstNamePublic",
          "isLastNamePublic",
          "isLocationPublic",
          "isBirthYearPublic",
          "isPreferredLanguagePublic",
          "currentPassword"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "firstName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "lastName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "location": {
            "maxLength": 200,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "birthYear": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "preferredLanguage": {
            "pattern": "^(fr|en)$",
            "type": "string"
          },
          "isFirstNamePublic": {
            "type": "boolean"
          },
          "isLastNamePublic": {
            "type": "boolean"
          },
          "isLocationPublic": {
            "type": "boolean"
          },
          "isBirthYearPublic": {
            "type": "boolean"
          },
          "isPreferredLanguagePublic": {
            "type": "boolean"
          },
          "currentPassword": {
            "maxLength": 128,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RegisterRequest": {
        "required": [
          "email",
          "username",
          "password",
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 0,
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 0,
            "type": "string"
          },
          "firstName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "lastName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "rateLimitAccount": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UserProfileResponse": {
        "required": [
          "id",
          "email",
          "emailVerified",
          "globalRole",
          "username",
          "firstName",
          "lastName",
          "location",
          "birthYear",
          "preferredLanguage",
          "isFirstNamePublic",
          "isLastNamePublic",
          "isLocationPublic",
          "isBirthYearPublic",
          "isPreferredLanguagePublic",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "globalRole": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "birthYear": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "preferredLanguage": {
            "type": "string"
          },
          "isFirstNamePublic": {
            "type": "boolean"
          },
          "isLastNamePublic": {
            "type": "boolean"
          },
          "isLocationPublic": {
            "type": "boolean"
          },
          "isBirthYearPublic": {
            "type": "boolean"
          },
          "isPreferredLanguagePublic": {
            "type": "boolean"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Instant"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Instant"
          }
        }
      }
    }
  },
...
}